third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/referrer-always-https-http.html

<!DOCTYPE html>
<html>
<head>
    <script src="/js-test-resources/js-test.js"></script>
    <script src="../resources/referrer-test.js"></script>
</head>
<body>
    <script>
        description("Test that 'always' works correctly from HTTPS to HTTP.");
        expectFullReferrer(ReferrerTest.ALWAYS, ReferrerTest.HTTPS, ReferrerTest.HTTP);
    </script>
</body>
</html>
