third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/referrer-always-https-http-expected.txt

Test that 'always' works correctly from HTTPS to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ReferrerTest.referrerResult is "https://127.0.0.1:8443/security/contentSecurityPolicy/resources/referrer-test.php?policy=always&to=http"
PASS successfullyParsed is true

TEST COMPLETE
